adding a settings page for courses
